move killscreen code, solve projectile bug
