+++ fighters/common/src/function_hooks/mod.rs
@@ -11,6 +11,7 @@ pub mod hitstun;
 pub mod controls;
 pub mod misc;
 pub mod jumps;
+pub mod killscreen;
 pub mod stage_hazards;
 pub mod set_fighter_status_data;
 pub mod attack;
@@ -741,6 +742,7 @@ pub fn install() {
     momentum_transfer::install();
     misc::install();
     jumps::install();
+    killscreen::install();
     stage_hazards::install();
     set_fighter_status_data::install();
     attack::install();
